requirements.txt
PyQt5==5.15.9
python-docx==0.8.11
docxtpl==0.16.8
kaleido=0.2.1
plotly=5.23.0
factur-x=3.1 
pypd=4.3.1

